Fix and test Table of Contents
